+++ CHANGELOG.md
@@ -10,6 +10,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Ensure rejected export wins are not included in yearly report
 
+### Removed
+
+- World bank tariff (data-store-service) pipeline
+- DIT BACI (data-store-service) pipeline
+
 ## 2020-04-21
 
 ### Changed